requirements.txt
carrier-client==0.2
certifi==2018.10.15
chardet==3.0.4
Django==2.1.2
django-carrier-client==0.2
djangorestframework==3.8.2
idna==2.7
jsonschema==2.6.0
psycopg2==2.7.5
psycopg2-binary==2.7.6.1
pytz==2018.7
requests==2.20.1
strict-rfc3339==0.7
urllib3==1.24.1
